+++ backend/src/main/java/org/camunda/optimize/service/es/reader/BackupWriter.java
@@ -35,7 +35,7 @@ public class BackupWriter {
   private final ConfigurationService configurationService;
   private final OptimizeIndexNameService indexNameService;
 
-  public void triggerSnapshotCreation(final Integer backupId) {
+  public void triggerSnapshotCreation(final Long backupId) {
     final String snapshot1Name = getSnapshotNameForImportIndices(backupId);
     final String snapshot2Name = getSnapshotNameForNonImportIndices(backupId);
     CompletableFuture.runAsync(() -> {
@@ -44,7 +44,7 @@ public class BackupWriter {
     });
   }
 
-  public void deleteOptimizeSnapshots(final Integer backupId) {
+  public void deleteOptimizeSnapshots(final Long backupId) {
     final DeleteSnapshotRequest deleteSnapshotRequest = new DeleteSnapshotRequest()
       .repository(configurationService.getEsSnapshotRepositoryName())
       .snapshots(getSnapshotPrefixWithBackupId(backupId) + "*");
@@ -109,7 +109,7 @@ public class BackupWriter {
     };
   }
 
-  private ActionListener<AcknowledgedResponse> getDeleteSnapshotActionListener(final Integer backupId) {
+  private ActionListener<AcknowledgedResponse> getDeleteSnapshotActionListener(final Long backupId) {
     return new ActionListener<>() {
       @Override
       public void onResponse(AcknowledgedResponse deleteSnapshotResponse) {
